move, remove action creator directories
